• 1.39f, 12 October 2007       □ Michiel Schmid contributed an improved Process>Subtract Background         command with new "Create Background", "Disable Smoothing" and         "Preview" options.       □ Added a "Transparent2" mode (zero pixels are transparent) to Edit>         Paste Control.       □ Maximized windows use more of the screen on OS X and Windows.       □ When shrinking selections, Edit>Selection>Enlarge now works with         selections touching the image edge.       □ File>Revert no longer deletes the selection.       □ Plugins>Compile and Run now runs compiled plugins (.class files)         without the Java compiler (usually in tools.jar) being present, and,         thanks to Johannes Schindelin, it uses com.sun.tools.javac instead of         the deprecated sun.tools.javac.       □ On Linux, the preferences file (IJ_Prefs.txt) is saved in the .imagej         directory and the menu font size is no longer set to 14 points.       □ The ImageCanvas.setDisplayList(Vector) method now accepts a list of         ShapeRois with differing colors and line widths (example).       □ Added the File.openAsRawString(path) macro function. Refer to the         First10Bytes and ZapGremlins macros for examples.       □ Added the COPY_ZERO_TRANSPARENT constant to the Blitter class.       □ Added the IJ.isMacro() method, which returns true if a macro is         running or an IJ.run() method is executing.       □ Added the getInstanceColor() and setStroke() methods to the Roi class.       □ Fixed a bug that could cause the imageCalculator() macro function to         return before the output window had opened.       □ Fixed a bug that could cause a boolean variable in a macro run() call         to trigger the inclusion of other boolean variables.       □ Fixed a bug that caused an exception if showProgress() was called in a         macro run from the command line using the -batch option.       □ Fixed a bug that could cause the image count to increase by one if the         setBatchMode(true) macro function was called twice in succession.

+++ ij/Prefs.java
@@ -100,12 +100,11 @@ public class Prefs {
 	public static String load(Object ij, Applet applet) {
 		InputStream f = ij.getClass().getResourceAsStream("/"+PROPS_NAME);
 		if (applet!=null)
-			return loadAppletProps(f,applet);
+			return loadAppletProps(f, applet);
 		if (homeDir==null)
 			homeDir = System.getProperty("user.dir");
 		String userHome = System.getProperty("user.home");
-		String osName = System.getProperty("os.name");
-		if (osName.indexOf("Windows",0)>-1) {
+		if (IJ.isWindows()) {
 			prefsDir = homeDir; //ImageJ folder on Windows
 			if (prefsDir.endsWith("Desktop"))
 				prefsDir = userHome;
@@ -113,6 +112,8 @@ public class Prefs {
 			prefsDir = userHome; // Mac Preferences folder or Unix home dir
 			if (IJ.isMacOSX())
 				prefsDir += "/Library/Preferences";
+			else
+				prefsDir += "/.imagej";
 		} 
 		if (f==null) {
 			try {f = new FileInputStream(homeDir+"/"+PROPS_NAME);}
@@ -256,11 +257,10 @@ public class Prefs {
 	static void loadPreferences() {
 		String path = prefsDir+separator+PREFS_NAME;
 		boolean ok =  loadPrefs(path);
-		if (!ok && IJ.isMacOSX()) {
+		if (!ok && !IJ.isWindows()) {
 			path = System.getProperty("user.home")+separator+PREFS_NAME;
 			ok = loadPrefs(path); // look in home dir
-			if (ok)
-				new File(path).delete();
+			if (ok) new File(path).delete();
 		}
 
 	}
@@ -304,6 +304,10 @@ public class Prefs {
 			GelAnalyzer.savePreferences(prefs);
 			NewImage.savePreferences(prefs);
 			String path = prefsDir+separator+PREFS_NAME;
+			if (prefsDir.endsWith(".imagej")) {
+				File f = new File(prefsDir);
+				if (!f.exists()) f.mkdir(); // create .imagej directory
+			}
 			savePrefs(prefs, path);
 		} catch (Throwable t) {
 			CharArrayWriter caw = new CharArrayWriter();
